pyproject.toml
[project]
name = "langgraph-sandbox"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ipykernel>=7.2.0",
    "langchain>=1.2.10",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.10",
    "langchain-text-splitters>=1.1.1",
    "langgraph>=1.0.9",
    "notebook>=7.5.3",
    "onnxruntime==1.16.3",
    "pypdf>=6.7.2",
    "pyppeteer>=2.0.0",
    "python-dotenv>=1.2.1",
]
